Bug Fix
Parameter estimation was reading the experiment files and stopping one
line short because of python's 0 index. This was fixed.

In the case of setting up a steady state experiment this meant copasi
was only reading the header line.

Additionally, when setting up a steady state the Role of the 0th column
was being given to time. This has been fixed. Setting up steady state
parameter estimation is now fully supported